+++ ena-submission/src/ena_deposition/create_raw_reads.py
@@ -29,9 +29,7 @@ from .notifications import SlackConfig, send_slack_notification, slack_conn_init
 from .submission_db_helper import (
     AccessionVersion,
     AssemblyTableEntry,
-    ProjectTableEntry,
     RawReadsTableEntry,
-    SampleTableEntry,
     Status,
     StatusAll,
     SubmissionTableEntry,
@@ -40,6 +38,7 @@ from .submission_db_helper import (
     find_conditions_in_db,
     find_errors_or_stuck_in_db,
     find_waiting_in_db,
+    get_project_and_sample_results,
     is_latest_revision,
     is_revision,
     previous_version,
@@ -48,33 +47,6 @@ from .submission_db_helper import (
 )
 
 logger = logging.getLogger(__name__)
-
-
-def get_address(config: Config, submission_row: SubmissionTableEntry) -> str | None:
-    if not config.is_broker:
-        return None
-    try:
-        group_details = call_loculus.get_group_info(config, submission_row.seq_metadata["groupId"])
-    except Exception as e:
-        logger.error(
-            f"Failed to fetch group info for groupId={submission_row.seq_metadata['groupId']}\n"
-            f"{traceback.format_exc()}"
-        )
-        msg = (
-            "Failed to fetch group info from Loculus for group: "
-            f"{submission_row.seq_metadata['groupId']}, {e}"
-        )
-        raise RuntimeError(msg) from e
-    address = group_details.address
-    address_list = [
-        submission_row.center_name,  # corresponds to Loculus' "Institution" group field
-        address.city,
-        address.state,
-        address.country,
-    ]
-    address_string = ", ".join([x for x in address_list if x])
-    logger.debug("Created address from group_info")
-    return address_string
 
 
 def get_assembly_values_in_metadata(config: Config, metadata: dict[str, str]) -> dict[str, str]:
@@ -137,7 +109,9 @@ def create_manifest_object(
             sample=sample_accession,
             description=get_description(config, metadata),
             **assembly_values,  # type: ignore
-            address=get_address(config, submission_row),
+            address=call_loculus.get_address(
+                config, submission_row.accession, submission_row.seq_metadata["groupId"]
+            ),
             fastq=submission_row.fastq_files,
         )
     except Exception as e:
@@ -378,39 +352,6 @@ def update_raw_reads_results_with_latest_version(db_engine: Engine, seq_key: Acc
         model_class=RawReadsTableEntry,
         reraise=False,
     )
-
-
-def get_project_and_sample_results(
-    db_engine: Engine, submission_row: SubmissionTableEntry
-) -> tuple[str, str]:
-    seq_key = {"accession": submission_row.accession, "version": submission_row.version}
-
-    sample_rows = find_conditions_in_db(db_engine, SampleTableEntry, conditions=seq_key)
-    if len(sample_rows) == 0:
-        error_msg = f"Entry {submission_row.accession} not found in sample_table"
-        raise RuntimeError(error_msg)
-
-    project_rows = find_conditions_in_db(
-        db_engine,
-        ProjectTableEntry,
-        conditions={"project_id": submission_row.project_id},
-    )
-    if len(project_rows) == 0:
-        error_msg = f"Entry {submission_row.accession} not found in project_table"
-        raise RuntimeError(error_msg)
-    sample_accession = (
-        sample_rows[0].result.get("ena_sample_accession") if sample_rows[0].result else None
-    )
-    study_accession = (
-        project_rows[0].result.get("bioproject_accession") if project_rows[0].result else None
-    )
-    if not sample_accession or not study_accession:
-        error_msg = (
-            f"Missing sample_accession or study_accession for accession {submission_row.accession} "
-            "cannot create manifest"
-        )
-        raise RuntimeError(error_msg)
-    return cast(str, sample_accession), cast(str, study_accession)
 
 
 def raw_reads_table_create(db_engine: Engine, config: Config):
